requirements.txt
psycopg2-binary~=2.9.10
pandas~=2.2.3
config~=0.5.1